backend/requirements.txt
flask
flask-restful
python-dotenv==0.12.0
Flask-PyMongo
